feat: add healthcheck api
