+++ package.json
@@ -23,7 +23,10 @@
     "ipfs-http-gateway": "^0.1.0",
     "ipfs-http-server": "^0.1.0",
     "multiformats": "^3.0.3",
-    "os-utils": "^0.0.14"
+    "os-utils": "^0.0.14",
+    "datastore-s3": "^3.0.0",
+    "aws-sdk": "^2.782.0",
+    "express": "^4.17.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.8",
@@ -32,7 +35,6 @@
     "eslint": "^6.8.0",
     "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-jest": "^24.1.0",
-    "express": "^4.17.1",
     "ts-node": "^8.8.1",
     "typescript": "^4.0.5"
   }
